Refact documents router to use request user
