+++ src/infra/http/utils/user-request-schema.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const userRequestSchema = z.object({
+  id: z.string(),
+  role: z.string(),
+});
